Add helper classes to create safe observables
In some rxjava classes if a subcription was disposed before
an onError happens, instead of silent catch the exception,
rxjava raises the exception, which causes a FC since the
exception was converted into an uncatched exception.

This patchse provides a SafeObservable.fromXXX methods
to create safe observables that doesn't raise uncatched
exceptions in case the subscription was disposed.